create folder

+++ controllers/artist.controller.js
@@ -1,4 +1,5 @@
 const Artists = require("../models/artist.model");
+const { uploadFile, deleteFileStream } = require("../config/s3");
 
 module.exports = {
   getAll: async (req, res) => {
@@ -32,9 +33,18 @@ module.exports = {
   addArtist: async (req, res) => {
     try {
       const data = req.body;
+
+      // upload file to aws s3
+      if (req.file) {
+        const result = await uploadFile(req.file);
+        console.log(result);
+        data.image = result.key;
+      }
+
       await Artists.create(data);
       res.json({
-        message: "Succes add Artist",
+        message: "Success add Artist",
+        data,
       });
     } catch (err) {
       console.log(err);
@@ -44,9 +54,21 @@ module.exports = {
 
   updateArtistById: async (req, res) => {
     try {
-      await Artists.updateOne({ _id: req.params.id }, { $set: req.body });
+      const artist = (await Artists.findById(req.params.id, "-__v")) || false;
+      const data = req.body;
+      // check file dan delete file lama
+      if (req.file) {
+        const result = await uploadFile(req.file);
+        console.log(result);
+        data.image = result.key;
+
+        // delete image on aws server
+        await deleteFileStream(artist.image);
+      }
+
+      await Artists.updateOne({ _id: req.params.id }, { $set: data });
       res.json({
-        message: "Succes Update Artist",
+        message: "Success Update Artist",
       });
     } catch (err) {
       console.log(err);
@@ -56,7 +78,11 @@ module.exports = {
 
   deleteArtistById: async (req, res) => {
     try {
+      const artist = (await Artists.findById(req.params.id, "-__v")) || false;
+      if (!artist) return res.status(404).json({ messege: "artist not found" });
+
       await Artists.deleteOne({ _id: req.params.id });
+      await deleteFileStream(artist.image);
       res.json({
         message: "Succes delete Artist",
       });
